support read in prow config as a dir
