+++ prow/cmd/hook/main.go
@@ -103,7 +103,7 @@ func main() {
 	logrus.SetFormatter(logrusutil.NewDefaultFieldsFormatter(nil, logrus.Fields{"component": "hook"}))
 
 	configAgent := &config.Agent{}
-	if err := configAgent.Start(o.configPath); err != nil {
+	if err := configAgent.Start(o.configPath, ""); err != nil {
 		logrus.WithError(err).Fatal("Error starting config agent.")
 	}
