+++ prow/cmd/horologium/main_test.go
@@ -116,7 +116,9 @@ func TestSync(t *testing.T) {
 	}
 	for _, tc := range testcases {
 		cfg := config.Config{
-			Periodics: []config.Periodic{{Name: "j"}},
+			JobConfig: config.JobConfig{
+				Periodics: []config.Periodic{{Name: "j"}},
+			},
 		}
 		cfg.Periodics[0].SetInterval(time.Minute)
 
@@ -181,7 +183,9 @@ func TestSyncCron(t *testing.T) {
 	}
 	for _, tc := range testcases {
 		cfg := config.Config{
-			Periodics: []config.Periodic{{Name: "j", Cron: "@every 1m"}},
+			JobConfig: config.JobConfig{
+				Periodics: []config.Periodic{{Name: "j", Cron: "@every 1m"}},
+			},
 		}
 
 		var jobs []kube.ProwJob
