+++ prow/cmd/sinker/main.go
@@ -54,7 +54,7 @@ func main() {
 	)
 
 	configAgent := &config.Agent{}
-	if err := configAgent.Start(*configPath); err != nil {
+	if err := configAgent.Start(*configPath, ""); err != nil {
 		logrus.WithError(err).Fatal("Error starting config agent.")
 	}
