+++ prow/cmd/tide/main.go
@@ -66,7 +66,7 @@ func main() {
 	)
 
 	configAgent := &config.Agent{}
-	if err := configAgent.Start(*configPath); err != nil {
+	if err := configAgent.Start(*configPath, ""); err != nil {
 		logrus.WithError(err).Fatal("Error starting config agent.")
 	}
