+++ prow/cmd/tot/main.go
@@ -282,7 +282,7 @@ func main() {
 		var configAgent *config.Agent
 		if o.configPath != "" {
 			configAgent = &config.Agent{}
-			if err := configAgent.Start(o.configPath); err != nil {
+			if err := configAgent.Start(o.configPath, ""); err != nil {
 				logrus.WithError(err).Fatal("Error starting config agent.")
 			}
 		}
