+++ prow/config/branch_protection_test.go
@@ -460,13 +460,14 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "undefined org not protected",
 			config: Config{
-				BranchProtection: BranchProtection{
-					Policy: Policy{
-						Protect: yes,
-					},
-					Orgs: map[string]Org{
-						"other": {},
-						// org not present
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						Policy: Policy{
+							Protect: yes,
+						},
+						Orgs: map[string]Org{
+							"unknown": {},
+						},
 					},
 				},
 			},
@@ -474,12 +475,14 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "protect via config default",
 			config: Config{
-				BranchProtection: BranchProtection{
-					Policy: Policy{
-						Protect: yes,
-					},
-					Orgs: map[string]Org{
-						"org": {},
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						Policy: Policy{
+							Protect: yes,
+						},
+						Orgs: map[string]Org{
+							"org": {},
+						},
 					},
 				},
 			},
@@ -488,11 +491,13 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "protect via org default",
 			config: Config{
-				BranchProtection: BranchProtection{
-					Orgs: map[string]Org{
-						"org": {
-							Policy: Policy{
-								Protect: yes,
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						Orgs: map[string]Org{
+							"org": {
+								Policy: Policy{
+									Protect: yes,
+								},
 							},
 						},
 					},
@@ -503,13 +508,15 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "protect via repo default",
 			config: Config{
-				BranchProtection: BranchProtection{
-					Orgs: map[string]Org{
-						"org": {
-							Repos: map[string]Repo{
-								"repo": {
-									Policy: Policy{
-										Protect: yes,
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						Orgs: map[string]Org{
+							"org": {
+								Repos: map[string]Repo{
+									"repo": {
+										Policy: Policy{
+											Protect: yes,
+										},
 									},
 								},
 							},
@@ -522,15 +529,17 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "protect specific branch",
 			config: Config{
-				BranchProtection: BranchProtection{
-					Orgs: map[string]Org{
-						"org": {
-							Repos: map[string]Repo{
-								"repo": {
-									Branches: map[string]Branch{
-										"branch": {
-											Policy: Policy{
-												Protect: yes,
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						Orgs: map[string]Org{
+							"org": {
+								Repos: map[string]Repo{
+									"repo": {
+										Branches: map[string]Branch{
+											"branch": {
+												Policy: Policy{
+													Protect: yes,
+												},
 											},
 										},
 									},
@@ -545,15 +554,17 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "ignore other org settings",
 			config: Config{
-				BranchProtection: BranchProtection{
-					Policy: Policy{
-						Protect: no,
-					},
-					Orgs: map[string]Org{
-						"other": {
-							Policy: Policy{Protect: yes},
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						Policy: Policy{
+							Protect: no,
 						},
-						"org": {},
+						Orgs: map[string]Org{
+							"other": {
+								Policy: Policy{Protect: yes},
+							},
+							"org": {},
+						},
 					},
 				},
 			},
@@ -562,13 +573,15 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "defined branches must make a protection decision",
 			config: Config{
-				BranchProtection: BranchProtection{
-					Orgs: map[string]Org{
-						"org": {
-							Repos: map[string]Repo{
-								"repo": {
-									Branches: map[string]Branch{
-										"branch": {},
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						Orgs: map[string]Org{
+							"org": {
+								Repos: map[string]Repo{
+									"repo": {
+										Branches: map[string]Branch{
+											"branch": {},
+										},
 									},
 								},
 							},
@@ -581,15 +594,17 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "pushers require protection",
 			config: Config{
-				BranchProtection: BranchProtection{
-					Policy: Policy{
-						Protect: no,
-						Restrictions: &Restrictions{
-							Teams: []string{"oncall"},
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						Policy: Policy{
+							Protect: no,
+							Restrictions: &Restrictions{
+								Teams: []string{"oncall"},
+							},
 						},
-					},
-					Orgs: map[string]Org{
-						"org": {},
+						Orgs: map[string]Org{
+							"org": {},
+						},
 					},
 				},
 			},
@@ -598,15 +613,17 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "required contexts require protection",
 			config: Config{
-				BranchProtection: BranchProtection{
-					Policy: Policy{
-						Protect: no,
-						RequiredStatusChecks: &ContextPolicy{
-							Contexts: []string{"test-foo"},
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						Policy: Policy{
+							Protect: no,
+							RequiredStatusChecks: &ContextPolicy{
+								Contexts: []string{"test-foo"},
+							},
 						},
-					},
-					Orgs: map[string]Org{
-						"org": {},
+						Orgs: map[string]Org{
+							"org": {},
+						},
 					},
 				},
 			},
@@ -615,18 +632,20 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "child policy with defined parent can disable protection",
 			config: Config{
-				BranchProtection: BranchProtection{
-					AllowDisabledPolicies: true,
-					Policy: Policy{
-						Protect: yes,
-						Restrictions: &Restrictions{
-							Teams: []string{"oncall"},
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						AllowDisabledPolicies: true,
+						Policy: Policy{
+							Protect: yes,
+							Restrictions: &Restrictions{
+								Teams: []string{"oncall"},
+							},
 						},
-					},
-					Orgs: map[string]Org{
-						"org": {
-							Policy: Policy{
-								Protect: no,
+						Orgs: map[string]Org{
+							"org": {
+								Policy: Policy{
+									Protect: no,
+								},
 							},
 						},
 					},
@@ -639,23 +658,27 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "Make required presubmits required",
 			config: Config{
-				BranchProtection: BranchProtection{
-					Policy: Policy{
-						Protect: yes,
-						RequiredStatusChecks: &ContextPolicy{
-							Contexts: []string{"cla"},
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						Policy: Policy{
+							Protect: yes,
+							RequiredStatusChecks: &ContextPolicy{
+								Contexts: []string{"cla"},
+							},
+						},
+						Orgs: map[string]Org{
+							"org": {},
 						},
 					},
-					Orgs: map[string]Org{
-						"org": {},
-					},
 				},
-				Presubmits: map[string][]Presubmit{
-					"org/repo": {
-						{
-							Name:      "required presubmit",
-							Context:   "required presubmit",
-							AlwaysRun: true,
+				JobConfig: JobConfig{
+					Presubmits: map[string][]Presubmit{
+						"org/repo": {
+							{
+								Name:      "required presubmit",
+								Context:   "required presubmit",
+								AlwaysRun: true,
+							},
 						},
 					},
 				},
@@ -670,18 +693,22 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "ProtectTested opts into protection",
 			config: Config{
-				BranchProtection: BranchProtection{
-					ProtectTested: true,
-					Orgs: map[string]Org{
-						"org": {},
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						ProtectTested: true,
+						Orgs: map[string]Org{
+							"org": {},
+						},
 					},
 				},
-				Presubmits: map[string][]Presubmit{
-					"org/repo": {
-						{
-							Name:      "required presubmit",
-							Context:   "required presubmit",
-							AlwaysRun: true,
+				JobConfig: JobConfig{
+					Presubmits: map[string][]Presubmit{
+						"org/repo": {
+							{
+								Name:      "required presubmit",
+								Context:   "required presubmit",
+								AlwaysRun: true,
+							},
 						},
 					},
 				},
@@ -696,20 +723,24 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "required presubmits require protection",
 			config: Config{
-				BranchProtection: BranchProtection{
-					Policy: Policy{
-						Protect: no,
-					},
-					Orgs: map[string]Org{
-						"org": {},
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						Policy: Policy{
+							Protect: no,
+						},
+						Orgs: map[string]Org{
+							"org": {},
+						},
 					},
 				},
-				Presubmits: map[string][]Presubmit{
-					"org/repo": {
-						{
-							Name:      "required presubmit",
-							Context:   "required presubmit",
-							AlwaysRun: true,
+				JobConfig: JobConfig{
+					Presubmits: map[string][]Presubmit{
+						"org/repo": {
+							{
+								Name:      "required presubmit",
+								Context:   "required presubmit",
+								AlwaysRun: true,
+							},
 						},
 					},
 				},
@@ -719,19 +750,23 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "Optional presubmits do not force protection",
 			config: Config{
-				BranchProtection: BranchProtection{
-					ProtectTested: true,
-					Orgs: map[string]Org{
-						"org": {},
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						ProtectTested: true,
+						Orgs: map[string]Org{
+							"org": {},
+						},
 					},
 				},
-				Presubmits: map[string][]Presubmit{
-					"org/repo": {
-						{
-							Name:      "optional presubmit",
-							Context:   "optional presubmit",
-							AlwaysRun: true,
-							Optional:  true,
+				JobConfig: JobConfig{
+					Presubmits: map[string][]Presubmit{
+						"org/repo": {
+							{
+								Name:      "optional presubmit",
+								Context:   "optional presubmit",
+								AlwaysRun: true,
+								Optional:  true,
+							},
 						},
 					},
 				},
@@ -740,23 +775,27 @@ func TestConfig_GetBranchProtection(t *testing.T) {
 		{
 			name: "Explicit configuration takes precedence over ProtectTested",
 			config: Config{
-				BranchProtection: BranchProtection{
-					ProtectTested: true,
-					Orgs: map[string]Org{
-						"org": {
-							Policy: Policy{
-								Protect: yes,
+				ProwConfig: ProwConfig{
+					BranchProtection: BranchProtection{
+						ProtectTested: true,
+						Orgs: map[string]Org{
+							"org": {
+								Policy: Policy{
+									Protect: yes,
+								},
 							},
 						},
 					},
 				},
-				Presubmits: map[string][]Presubmit{
-					"org/repo": {
-						{
-							Name:      "optional presubmit",
-							Context:   "optional presubmit",
-							AlwaysRun: true,
-							Optional:  true,
+				JobConfig: JobConfig{
+					Presubmits: map[string][]Presubmit{
+						"org/repo": {
+							{
+								Name:      "optional presubmit",
+								Context:   "optional presubmit",
+								AlwaysRun: true,
+								Optional:  true,
+							},
 						},
 					},
 				},
