+++ testgrid/cmd/configurator/config_test.go
@@ -346,7 +346,7 @@ func TestJobsTestgridEntryMatch(t *testing.T) {
 
 	jobs := make(map[string]bool)
 
-	prowConfig, err := prow_config.Load(prowPath)
+	prowConfig, err := prow_config.Load(prowPath, "")
 	if err != nil {
 		t.Fatalf("Could not load prow configs: %v\n", err)
 	}
